requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.2
flake8==3.9.2
gunicorn==20.1.0
mccabe==0.6.1
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
